+++ src/attacks/evasion_attacks.py
@@ -1,0 +1,13 @@
+from attacks.attack_base import Attacker
+
+
+class EvasionAttacker(Attacker):
+    def __init__(self, **kwargs):
+        super().__init__()
+
+
+class EmptyEvasionAttacker(EvasionAttacker):
+    name = "EmptyEvasionAttacker"
+
+    def attack(self, **kwargs):
+        pass
